+++ Cargo.toml
@@ -13,15 +13,8 @@ categories = ["development-tools", "parser-implementations", "parsing", "simulat
 
 [[bin]]
 path = "rezasm-app/rezasm-cli/src/main.rs"
-name = "rezasm-cli"
-
-[[bin]]
-path = "rezasm-app/rezasm-tauri/src/main.rs"
-name = "rezasm-tauri"
-
-[lib]
-path = "rezasm-app/rezasm-wasm/src/lib.rs"
-name = "rezasm_wasm"
+name = "rezasm-cli-app"
+test = false
 
 
 [build-dependencies]
@@ -51,6 +44,10 @@ thiserror = "1.0.47"
 wasm-bindgen = "0.2.87"
 
 [workspace]
-members = ["rezasm-app/rezasm-cli", "rezasm-app/rezasm-tauri", "rezasm-app/rezasm-wasm", 
-    "rezasm-source/rezasm-core", "rezasm-source/rezasm-web-core",]
-default-members = ["rezasm-app/rezasm-cli", "rezasm-source/rezasm-core"]
+members = [
+    "rezasm-app/rezasm-cli",
+    "rezasm-app/rezasm-tauri",
+    "rezasm-app/rezasm-wasm",
+    "rezasm-source/rezasm-core",
+    "rezasm-source/rezasm-web-core",
+]
